Created new package for qr3 and refactored tests to be run with unittest.
