+++ qr3/__init__.py
@@ -1,0 +1,1 @@
+"""This package contains the qr classes."""
