+++ qr3/qr.py
